[misc] rapidjson -> nlohmann_json
Rationale.
- Prefer usability/maintainability over raw performances.

RapidJSON optimizes the memory usage and the objects cannot be
copied directly. This means that using RapidJSON in our case
implies non-trivial memory management techniques, which is not desired.

More generally, the RapidJSON API is designed to allow good performance,
not to be easy-to-use. I think that the netorcai-client-cpp project
does not need critical performance, and that focusing on its
maintainability would be more adequate.